% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{award}
\alias{award}
\title{Example data: Awards}
\format{
\subsection{\code{award}}{

A dataframe of awards
\describe{
\item{\code{title}}{Title of award}
\item{\code{location}}{Location of award}
\item{\code{date}}{Date of award}
\item{\code{description}}{Description of award}
}
}
}
\source{
data-raw/data.R
}
\usage{
award
}
\description{
Example data: Awards
}
\keyword{datasets}
