% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ}
\alias{educ}
\title{Example data: Education}
\format{
\subsection{\code{educ}}{

A dataframe of educational history
\describe{
\item{\code{title}}{Degree title}
\item{\code{location}}{Location of institution}
\item{\code{date}}{Date of degree}
\item{\code{description}}{Institution}
}
}
}
\source{
data-raw/data.R
}
\usage{
educ
}
\description{
Example data: Education
}
\keyword{datasets}
