% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date}
\alias{format_date}
\title{Data Column Formatter}
\usage{
format_date(
  data,
  start = "start",
  end = NULL,
  sep = " - ",
  date_format = "\%b \%Y",
  colname_date = "date",
  replace_na = "Present",
  sort_by = "none",
  decreasing = TRUE
)
}
\arguments{
\item{data}{Data frame for resume entry}

\item{start}{Start date column name. Default is "start"}

\item{end}{Optional. End date column name.}

\item{sep}{Separator between start and end date. Default is " - "}

\item{date_format}{Date format. Default is "\%B \%Y"}

\item{colname_date}{Name for the generated date column. Default is "date"}

\item{replace_na}{Value to replace NA values. Default is "Present"}

\item{sort_by}{Sort by "none", "start" or "end". Default is "none"}

\item{decreasing}{Sort in decreasing order. Default is TRUE}
}
\value{
Data frame with formatted date column
}
\description{
Data Column Formatter
}
\examples{
work |>
  format_date(end = "end", date_format = "\%Y", sort_by = "start") |>
  resume_entry()

}
