% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub}
\alias{pub}
\title{Example data: Publications}
\format{
\subsection{\code{pub}}{

A dataframe of publications
\describe{
\item{\code{title}}{Title of publication}
\item{\code{location}}{Journal}
\item{\code{date}}{Date of publication}
}
}
}
\source{
data-raw/data.R
}
\usage{
pub
}
\description{
Example data: Publications
}
\keyword{datasets}
