% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typstcv.R
\name{resume_entry}
\alias{resume_entry}
\title{Convert a data frame to a resume entry}
\usage{
resume_entry(
  data,
  title = "title",
  location = "location",
  date = "date",
  description = "description",
  details = NULL
)
}
\arguments{
\item{data}{A data frame with columns for title, location, date, and description}

\item{title}{The column name for the title. Default is "title"}

\item{location}{The column name for the location. Default is "location"}

\item{date}{The column name for the date. Default is "date"}

\item{description}{The column name for the description. Default is "description"}

\item{details}{The column name for additional details. Default is NULL}
}
\value{
A string with the resume entry
}
\description{
Convert a data frame to a resume entry
}
\examples{
educ <- data.frame(
 title = c("Ph.D. in Physics", "Master of Science"),
 location = c("Zürich, Switzerland", "Zürich, Switzerland"),
 date = c("1905", "1896 - 1900"),
 description = c("University of Zürich", "ETH")
)

resume_entry(educ)

# With details
award <- data.frame(
 title = c("Nobel Prize in Physics"),
 location = c("Stockholm, Sweden"),
 date = c("1921"),
 description = c("For his services to"),
 detail1 = c("Theoretical Physics"),
 detail2 = c("Discovery of the law of the photoelectric effect")
)

resume_entry(award, details = c("detail1", "detail2"))

}
