% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skill}
\alias{skill}
\title{Example data: Skills}
\format{
\subsection{\code{skill}}{

A dataframe of skills
\describe{
\item{\code{title}}{Category of skill}
\item{\code{description}}{List of skills}
}
}
}
\source{
data-raw/data.R
}
\usage{
skill
}
\description{
Example data: Skills
}
\keyword{datasets}
