% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{work}
\alias{work}
\title{Example data: Work experience}
\format{
\subsection{\code{work}}{

A dataframe of work experience
\describe{
\item{\code{title}}{Job title}
\item{\code{location}}{Location of job}
\item{\code{date}}{Date of employment}
\item{\code{description}}{Employer}
}
}
}
\source{
data-raw/data.R
}
\usage{
work
}
\description{
Example data: Work experience
}
\keyword{datasets}
